Use FFmpeg to transcode all audio files (removes Sox)
