+++ lib/mix/tasks/ask.convert_wav_files_to_mp3_in_db.ex
@@ -1,7 +1,7 @@
 defmodule Mix.Tasks.Ask.ConvertWavFilesToMp3InDb do
   use Mix.Task
 
-  alias Ask.{Sox, Audio, Repo}
+  alias Ask.{FFmpeg, Audio, Repo}
 
   import Ecto.Query
 
@@ -14,7 +14,7 @@ defmodule Mix.Tasks.Ask.ConvertWavFilesToMp3InDb do
     File.write(path, audio.data, [:binary])
     mp3_filename = "#{Path.basename(audio.filename, ".wav")}.mp3"
 
-    case Sox.convert(path, "mp3") do
+    case FFmpeg.convert(path, "mp3") do
       {:ok, mp3} ->
         case Repo.update(Audio.changeset(audio, %{"data" => mp3, "filename" => mp3_filename})) do
           {:ok, _} ->
